// src/app/layout/candidate/model/Candidate.ts
export class Candidate {
    candidateId: string;
    candidateName: string;
    candidateEmail: string;
    candidateContact: string;
    responseStatus:string;
    interviewId:string;
    startTime:string;
    endTime:string;   
    practice: string;
    customerAC: string;
    quarter:string;
    notes:string;
    offerPU:string;
    demandId:string;   
    location: string;
    doj: string;
    status: any;
    remark:string;
    background:string;
    weekEndingDate:string;
    recievedMonth:string;   
    parentSkillSet: string;
    responsible: string;
    profileAge: string;
    firstPanelistEmailId: string;
    secondPanelistEmailId:string;

    // intialStatusValue:any
}